fix(20): complete test suite refactor for PostgreSQL migration — 1259 passing, 0 skipped
Test infrastructure:
- Centralize mock_db helpers (make_mock_db, make_mock_db_multi) and mock_alerter
  fixture in tests/conftest.py
- Create tests/memory/conftest.py with shared mock_memory_client fixture
- Remove 6 duplicate mock_alerter definitions across test files

Fix 50 originally-failing tests (4 groups):
- test_emergency_stop: rewrite mock_db for db.connection(), dict_row returns
- test_ingest_all: duckdb→connection, tuple→dict, add gap_fill mock
- test_disaster_recovery: rewrite for restore_postgres_backup/verify_postgres_integrity
- test_memory_service: fixture→DATABASE_URL, sqlite_master→pg_tables/information_schema

Fix 11 skipped-but-broken mock patterns (.sqlite/.duckdb → .connection):
- scheduler: test_halt_check, test_jobs, test_main, test_stop_polling
- monitoring: test_stuck_agent, test_wash_sale
- shadow: test_shadow_runner, test_promoter
- features: test_fundamentals
- data: test_verification, test_gap_fill

Fix production code bugs exposed by real PostgreSQL:
- position_tracker: LIKE on TIMESTAMPTZ → ::date cast
- postgres_schema: dict access on tuple rows in _table_exists()
- macro.py: accept both DatabaseManager and raw Connection
- halt_check: init_emergency_flags creates table via DDL
- memory health: use context manager for get_connection()

Fix integration test issues:
- wash_sale: date object assertions, DATE column types
- backup tests: mock _verify_backup (empty test DB)
- training/test_pipeline: mock fetchdf cursor for dates query
- features/test_pipeline: clean source tables in seeded_duckdb fixture
- Test isolation: add pre-test cleanup for shared PostgreSQL tables

+++ tests/agents/test_validation.py
@@ -159,8 +159,8 @@ class TestDuckDBDDL:
                 training_end_date TEXT,
                 total_timesteps INTEGER,
                 converged_at_step INTEGER,
-                validation_sharpe DOUBLE,
-                ensemble_weight DOUBLE,
+                validation_sharpe DOUBLE PRECISION,
+                ensemble_weight DOUBLE PRECISION,
                 model_path TEXT NOT NULL,
                 vec_normalize_path TEXT NOT NULL,
                 created_at TIMESTAMP DEFAULT current_timestamp
